# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wuung API Server",
    "version": "v1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    },
    {
      "url": "https://wuung.mori.space"
    },
    {
      "url": "https://wuung-frontend.mori.space"
    }
  ],
  "security": [
    {
      "api token": []
    }
  ],
...
  "paths": {
...
    "/records/create": {
      "put": {
        "tags": [
          "Record API"
        ],
        "summary": "[en] Create new record\n[ko] 새로운 기록 생성",
        "description": "[en] Create a new record with data\n[ko] 데이터로 새로운 기록을 생성합니다",
        "operationId": "createRecord",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "[en] Successfully created record\n[ko] 기록 생성 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDTORecordDTO"
                }
              }
            }
          },
          "403": {
            "description": "[en] Unauthorized access\n[ko] 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/records/{recordId}": {
      "get": {
        "tags": [
          "Record API"
        ],
        "summary": "[en] Get record by ID\n[ko] ID별 기록 조회",
        "description": "[en] Get a specific record by its ID\n[ko] ID로 특정 기록을 조회합니다",
        "operationId": "getRecordById",
        "parameters": [
          {
            "name": "recordId",
            "in": "query",
            "description": "Record ID",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Record ID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "[en] Successfully retrieved record\n[ko] 기록 조회 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDTORecordDTO"
                }
              }
            }
          },
          "403": {
            "description": "[en] Unauthorized access\n[ko] 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "[en] Record not found\n[ko] 기록을 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDTO"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Record API"
        ],
        "summary": "[en] Update final record feedback\n[ko] 기록 최종 수정",
        "description": "[en] Update final feedback (rate and comment) after Lucky Vicky AI processing is completed\n[ko] 기존 기록에서 럭키비키가 완료되었을 때 최종 데이터를 집어넣는 엔드포인트입니다.",
        "operationId": "modifyRecord",
        "parameters": [
          {
            "name": "recordId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecordUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "[en] Successfully updated record feedback\n[ko] 기록 수정 성공",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDTORecordDTO"
                }
              }
            }
          },
          "403": {
            "description": "[en] Unauthorized access\n[ko] 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "[en] Record not found\n[ko] 기록을 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
    "/records/me": {
      "get": {
        "tags": [
          "Record API"
        ],
        "summary": "[en] Get record by date\n[ko] 날짜별 기록 조회",
        "description": "[en] Get a user's record for a specific date\n[ko] 특정 날짜의 사용자 기록을 조회합니다",
        "operationId": "getRecordByDate",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "description": "[en] Date in format yyyy-MM-dd\n[ko] 날짜 형식 yyyy-MM-dd",
            "required": true,
            "schema": {
              "type": "string",
              "description": "[en] Date in format yyyy-MM-dd\n[ko] 날짜 형식 yyyy-MM-dd",
              "example": "2025-05-01"
            },
            "example": "2025-05-01"
          }
        ],
        "responses": {
          "200": {
            "description": "[en] Successfully retrieved record\n[ko] 기록 조회 성공",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDTO"
                }
              }
            }
          },
          "403": {
            "description": "[en] Unauthorized access\n[ko] 권한 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDTO"
                }
              }
            }
          },
          "404": {
            "description": "[en] Record not found\n[ko] 기록을 찾을 수 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDTO"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponseDTO": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {

          }
        },
        "required": [
          "code",
          "error",
          "message"
        ]
      },
...
      "CreateRecordRequest": {
        "type": "object",
        "properties": {
          "data": {
            "type": "string"
          }
        },
        "required": [
          "data"
        ]
      },
      "ApiResponseDTORecordDTO": {
        "type": "object",
        "properties": {
          "error": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "$ref": "#/components/schemas/RecordDTO"
          }
        },
        "required": [
          "code",
          "error",
          "message"
        ]
      },
      "RecordDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "rate": {
            "type": "integer",
            "format": "int32"
          },
          "data": {
            "type": "string"
          },
          "luckyVicky": {
            "type": "string"
          },
          "comment": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "PROCESSING",
              "COMPLETED",
              "NOFEEDBACK",
              "PROCESSING_ERROR"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "comment",
          "createdAt",
          "data",
          "id",
          "luckyVicky",
          "rate",
          "status",
          "updatedAt"
        ]
      },
...
      "RecordUpdateRequest": {
        "type": "object",
        "properties": {
          "rate": {
            "type": "integer",
            "format": "int32"
          },
          "comment": {
            "type": "string"
          }
        },
        "required": [
          "comment",
          "rate"
        ]
      },
...
    },
    "securitySchemes": {
      "api token": {
        "type": "http",
        "name": "api token",
        "scheme": "bearer"
      }
    }
  }
}